New: manpages, html pages and packaging

+++ install.sh
@@ -2,6 +2,8 @@
 #easier way how to handle version formats http://safrm.net/projects/appver/
 #author:  Miroslav Safr <miroslav.safr@gmail.com>
 BINDIR=/usr/bin/
+DOCDIR=/usr/share/doc
+MANDIR=/usr/share/man
 
 #root check
 USERID=`id -u`
@@ -10,7 +12,24 @@ USERID=`id -u`
     exit 0
 }
 
+#test
+sh -n ./appver
+if  [ $? != 0 ]; then
+	echo "syntax error in appver, exiting.." 
+	exit 1
+fi
+
 . ./appver
 install -m 0777 -v ./appver  $BINDIR
 sed -i".bkp" "1,/^VERSION=/s/^VERSION=.*/VERSION=$APP_FULL_VERSION_TAG/" $BINDIR/appver && rm -f $BINDIR/appver.bkp
 sed -i".bkp" "1,/^VERSION_DATE=/s/^VERSION_DATE=.*/VERSION_DATE=$APP_BUILD_DATE/" $BINDIR/appver && rm -f $BINDIR/appver.bkp
+
+MANPAGES=`find ./doc/manpages -type f`
+install -d -m 755 $MANDIR/man1
+install -m 644 $MANPAGES $MANDIR/man1
+
+DOCS="./README ./LICENSE.LGPL"
+install -d -m 755 $DOCDIR/appver
+install -m 644 $DOCS $DOCDIR/appver
+
+
